[refactor] import location 수정
